Feature/#161

+++ service/src/main/java/way/application/service/bookMark/dto/response/BookMarkResponseDto.java
@@ -27,7 +27,7 @@ public class BookMarkResponseDto {
 	public record BookMarkImageInfo(
 		Long feedImageSeq,
 		String feedImageURL,
-		Long feedImageOrder
+		Integer feedImageOrder
 	) {
 
 	}
